+++ Library/Account.cs
@@ -41,6 +41,7 @@ namespace Recurly
         public string HostedLoginToken { get; private set; }
         public DateTime CreatedAt { get; private set; }
         public DateTime UpdatedAt { get; private set; }
+        public DateTime? ClosedAt { get; private set; }
         public bool VatLocationValid { get; private set; }
         public Address Address { get; set; }
         public bool HasLiveSubscription { get; private set; }
@@ -48,6 +49,32 @@ namespace Recurly
         public bool HasFutureSubscription { get; private set; }
         public bool HasCanceledSubscription { get; private set; }
         public bool HasPastDueInvoice { get; private set; }
+        public string PreferredLocale { get; set; }
+
+        private AccountAcquisition _accountAcquisition;
+
+        public AccountAcquisition AccountAcquisition
+        {
+            get
+            {
+                if (_accountAcquisition != null)
+                    return _accountAcquisition;
+
+                try
+                {
+                    _accountAcquisition = AccountAcquisition.Get(AccountCode);
+                }
+                catch (NotFoundException)
+                {
+                    _accountAcquisition = null;
+                }
+                return _accountAcquisition;
+            }
+            set
+            {
+                _accountAcquisition = value;
+            }
+        }
 
         private BillingInfo _billingInfo;
 
@@ -140,6 +167,16 @@ namespace Recurly
         { }
 
         /// <summary>
+        /// Remove an account's acquisition data.
+        /// </summary>
+        public void DeleteAccountAcquisition()
+        {
+            Client.Instance.PerformRequest(Client.HttpRequestMethod.Delete,
+                UrlPrefix + Uri.EscapeDataString(AccountCode) + "/acquisition");
+            _accountAcquisition = null;
+        }
+
+        /// <summary>
         /// Delete an account's billing info.
         /// </summary>
         public void DeleteBillingInfo()
@@ -195,29 +232,31 @@ namespace Recurly
         /// <summary>
         /// Posts pending charges on an account
         /// </summary>
-        public Invoice InvoicePendingCharges(Invoice invoice = null)
+        public InvoiceCollection InvoicePendingCharges(Invoice invoice = null)
         {
-            var i = invoice ?? new Invoice();
+            invoice = invoice ?? new Invoice();
+            var collection = new InvoiceCollection();
             Client.Instance.PerformRequest(Client.HttpRequestMethod.Post,
                 UrlPrefix + Uri.EscapeDataString(AccountCode) + "/invoices",
-                i.WriteXml,
-                i.ReadXml);
+                invoice.WriteXml,
+                collection.ReadXml);
 
-            return i;
+            return collection;
         }
 
         /// <summary>
         /// Previews a new invoice for the pending charges on an account
         /// </summary>
-        public Invoice PreviewInvoicePendingCharges(Invoice invoice = null)
+        public InvoiceCollection PreviewInvoicePendingCharges(Invoice invoice = null)
         {
-            var i = invoice ?? new Invoice();
+            invoice = invoice ?? new Invoice(); 
+            var collection = new InvoiceCollection();
             Client.Instance.PerformRequest(Client.HttpRequestMethod.Post,
                 UrlPrefix + Uri.EscapeDataString(AccountCode) + "/invoices/preview",
-                i.WriteXml,
-                i.ReadXml);
+                invoice.WriteXml,
+                collection.ReadXml);
 
-            return i;
+            return collection;
         }
 
         /// <summary>
@@ -238,6 +277,37 @@ namespace Recurly
 
             return statusCode == HttpStatusCode.NotFound ? null : adjustments;
         }
+
+        /// <summary>
+        /// Gets all adjustments for this account
+        /// </summary>
+        /// <param name="filter">Optional filter criteria</param>
+        /// <param name="type">Adjustment type to retrieve. Optional, default: All.</param>
+        /// <param name="state">State of the Adjustments to retrieve. Optional, default: Any.</param>
+        /// <returns></returns>
+        public RecurlyList<Adjustment> GetAdjustments(FilterCriteria filter, Adjustment.AdjustmentType type = Adjustment.AdjustmentType.All,
+            Adjustment.AdjustmentState state = Adjustment.AdjustmentState.Any)
+        {
+            var adjustments = new AdjustmentList();
+            filter = filter ?? FilterCriteria.Instance;
+            var parameters = filter.ToNamedValueCollection();
+            if (type != Adjustment.AdjustmentType.All) {
+                parameters["type"] = type.ToString().EnumNameToTransportCase();
+            }
+            if (state != Adjustment.AdjustmentState.Any)
+            {
+                parameters["state"] = state.ToString().EnumNameToTransportCase();
+            }
+
+
+            var statusCode = Client.Instance.PerformRequest(Client.HttpRequestMethod.Get,
+                UrlPrefix + Uri.EscapeDataString(AccountCode) + "/adjustments/"
+                + "?" + parameters.ToString()
+                , adjustments.ReadXmlList);
+
+            return statusCode == HttpStatusCode.NotFound ? null : adjustments;
+        }
+
 
         /// <summary>
         /// Gets all shipping addresses
@@ -366,6 +436,8 @@ namespace Recurly
 
                 if (reader.NodeType != XmlNodeType.Element) continue;
 
+                DateTime dt;
+
                 switch (reader.Name)
                 {
                     case "account_code":
@@ -421,6 +493,11 @@ namespace Recurly
                         HostedLoginToken = reader.ReadElementContentAsString();
                         break;
 
+                    case "closed_at":
+                        if (DateTime.TryParse(reader.ReadElementContentAsString(), out dt))
+                            ClosedAt = dt;
+                        break;
+
                     case "created_at":
                         CreatedAt = reader.ReadElementContentAsDateTime();
                         break;
@@ -469,6 +546,10 @@ namespace Recurly
                         if (bool.TryParse(reader.ReadElementContentAsString(), out e))
                             HasPastDueInvoice = e;
                         break;
+
+                    case "preferred_locale":
+                        PreferredLocale = reader.ReadElementContentAsString();
+                        break;
                 }
             }
         }
@@ -492,11 +573,15 @@ namespace Recurly
             xmlWriter.WriteStringIfValid("vat_number", VatNumber);
             xmlWriter.WriteStringIfValid("entity_use_code", EntityUseCode);
             xmlWriter.WriteStringIfValid("cc_emails", CcEmails);
+            xmlWriter.WriteStringIfValid("preferred_locale", PreferredLocale);
 
             xmlWriter.WriteIfCollectionHasAny("shipping_addresses", ShippingAddresses);
 
             if (TaxExempt.HasValue)
                 xmlWriter.WriteElementString("tax_exempt", TaxExempt.Value.AsString());
+
+            if(_accountAcquisition != null)
+                _accountAcquisition.WriteXml(xmlWriter);
 
             if (_billingInfo != null)
                 _billingInfo.WriteXml(xmlWriter);